A1/ecseX46-submission-master/requirements.txt
numpy == 1.14.5
OpenEXR == 1.3.0
Pillow == 5.1.0
matplotlib == 2.2.2
beautifulsoup4 == 4.6.0
html5lib == 1.0.1
